Some simple fixes and add lookup on user list
